inference/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
Pillow==11.1.0
numpy==2.2.3
torch==2.5.1
transformers==4.48.3
pydantic==2.10.5
